Move userbar to separate component
