+++ src/components/UserChip/index.ts
@@ -1,0 +1,2 @@
+export * from "./UserChip";
+export { default } from "./UserChip";
